requirements.txt
numpy
scipy
tk
pillow
sounddevice
soundfile
pydub
google-cloud-texttospeech
google-cloud-speech
openai
pygame